Fixed a bug minor mistake. Now using Double.Parse instead of Int32 so that decimals will work

+++ OOP_Calculator_JACOB/Form1.cs
@@ -99,7 +99,7 @@ namespace OOP_Calculator_JACOB
                 default:
                     break;
             }
-            result_Value = Int32.Parse(textBox1.Text);
+            result_Value = Double.Parse(textBox1.Text);
             operation_Performed = "";
 
             equal_Check = true;
